pyproject.toml
[tool.poetry]
name = "pagepy"
version = "0.1.0"
description = "simple & opinionated static site gen."
authors = ["Lasse Maywald"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.16"
jinja2 = "^3.1.2"
markdown = "^3.4.1"
python-frontmatter = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
